Resolve some file misplacement issues
