+++ server/server/views/category.py
@@ -2,14 +2,13 @@
 from cornice.resource import resource, view
 from pyramid.security import Allow, Authenticated
 
-from ...models import model_to_dict
-from ...models.event_tag import EventTag
-from ...models.tag import Tag
+from ..models import model_to_dict
+from ..models.category import Category
 
 
-@resource(collection_path='/tag', path='/tag/{tag_id}',
+@resource(collection_path='/category', path='/category/{category_id}',
           renderer='json', cors_origins=('http://localhost:3000',))
-class TagView(object):
+class CategoryView(object):
 
     def __init__(self, request, context=None):
         self.request = request
@@ -22,9 +21,8 @@ class TagView(object):
     def collection_get(self):
         request = self.request
         response = {
-            'success': False,
-            'tags': None
+            'categories': None
         }
-        tags = Tag.get_all(request)
-        response['tags'] = [obj.tag for obj in tags]
+        categories = Category.get_all(request)
+        response['categories'] = [obj.category for obj in categories]
         return response
