+++ server/server/views/event.py
@@ -7,16 +7,16 @@ from cornice.validators import colander_body_validator
 from pyramid.security import Allow, Authenticated, Everyone
 from pyramid.httpexceptions import HTTPNotFound
 
-from ...models import model_to_dict
-from ...models.category import Category
-from ...models.event import Event
-from ...models.event_history import EventHistory
-from ...models.event_status import EventStatus
-from ...models.event_tag import EventTag
-from ...models.tag import Tag
-from ...models.subscribe import Subscribe
-from ...validation_schema import EventSchema
-from ...models.user import User
+from ..models import model_to_dict
+from ..models.category import Category
+from ..models.event import Event
+from ..models.event_history import EventHistory
+from ..models.event_status import EventStatus
+from ..models.event_tag import EventTag
+from ..models.tag import Tag
+from ..models.subscribe import Subscribe
+from ..validation_schema import EventSchema
+from ..models.user import User
 
 
 @resource(collection_path='/event', path='/event/{event_id}', renderer='json',
