+++ server/server/views/registration.py
